requirements.txt
streamlit
python-dotenv
pydantic
email-validator
groq
pdfplumber
python-docx